refactor packages from org.acme to com.scanales
