+++ quarkus-app/pom.xml
@@ -2,7 +2,7 @@
 <project>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.acme</groupId>
+    <groupId>com.scanales</groupId>
     <artifactId>getting-started</artifactId>
     <version>1.0.0-SNAPSHOT</version>
