+++ quarkus-app/src/main/java/com/scanales/eventflow/public_/LoginPage.java
@@ -1,0 +1,26 @@
+package com.scanales.eventflow.public_;
+
+import io.quarkus.qute.CheckedTemplate;
+import io.quarkus.qute.TemplateInstance;
+
+import jakarta.annotation.security.PermitAll;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+
+@Path("/login")
+public class LoginPage {
+
+    @CheckedTemplate
+    static class Templates {
+        static native TemplateInstance login();
+    }
+
+    @GET
+    @PermitAll
+    @Produces(MediaType.TEXT_HTML)
+    public TemplateInstance login() {
+        return Templates.login();
+    }
+}
