+++ quarkus-app/src/main/java/com/scanales/getting/started/GreetingService.java
@@ -1,0 +1,12 @@
+package com.scanales.getting.started;
+
+import jakarta.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class GreetingService {
+
+    public String greeting(String name) {
+        return "hello " + name;
+    }
+
+}
