--- quarkus-app/src/main/java/org/acme/getting/started/GreetingService.java
@@ -1,12 +1,0 @@
-package org.acme.getting.started;
-
-import jakarta.enterprise.context.ApplicationScoped;
-
-@ApplicationScoped
-public class GreetingService {
-
-    public String greeting(String name) {
-        return "hello " + name;
-    }
-
-}
