+++ quarkus-app/src/test/java/com/scanales/getting/started/GreetingResourceIT.java
@@ -1,0 +1,9 @@
+package com.scanales.getting.started;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class GreetingResourceIT extends GreetingResourceTest {
+
+    // Execute the same tests but in native mode.
+}
