+++ quarkus-app/src/test/java/com/scanales/logout/LogoutResourceIT.java
@@ -1,0 +1,8 @@
+package com.scanales.logout;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class LogoutResourceIT extends LogoutResourceTest {
+    // Runs the same tests in native mode
+}
