--- quarkus-app/src/test/java/org/acme/getting/started/GreetingResourceIT.java
@@ -1,9 +1,0 @@
-package org.acme.getting.started;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class GreetingResourceIT extends GreetingResourceTest {
-
-    // Execute the same tests but in native mode.
-}
