--- quarkus-app/src/test/java/org/acme/getting/started/GreetingResourceTest.java
@@ -1,35 +1,0 @@
-package org.acme.getting.started;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-
-import java.util.UUID;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-
-@QuarkusTest
-public class GreetingResourceTest {
-
-    @Test
-    public void testHelloEndpoint() {
-        given()
-                .when().get("/hello")
-                .then()
-                .statusCode(200)
-                .body(is("hello"));
-    }
-
-    @Test
-    public void testGreetingEndpoint() {
-        String uuid = UUID.randomUUID().toString();
-        given()
-                .pathParam("name", uuid)
-                .when().get("/hello/greeting/{name}")
-                .then()
-                .statusCode(200)
-                .body(is("hello " + uuid));
-    }
-
-}
