--- quarkus-app/src/test/java/org/acme/logout/LogoutResourceIT.java
@@ -1,8 +1,0 @@
-package org.acme.logout;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class LogoutResourceIT extends LogoutResourceTest {
-    // Runs the same tests in native mode
-}
